Bump to version 1.0.0-rc14
Should be close to a final 1.0.0

A few potential breaking changes.

The oauth models have been refactored to be simple maps instead of
records. The factory functions are simpler and take a simple map of
data.

If you have your own token and authcode creator functions for versions
prior to 1.0.0-rc14 please verify that they work when upgrading.